+++ src/lib.rs
@@ -1,9 +1,10 @@
 pub mod error;
 mod idl;
-pub mod ser;
+mod loader;
 mod value;
 
 pub use idl::from_str;
+pub use idl::from_str_flat;
 pub use value::*;
 
 #[cfg(feature = "binary")]
@@ -32,14 +33,14 @@ pub fn encode(input: &[Value]) -> Vec<u8> {
 
 #[cfg(test)]
 mod test {
-    use crate::{decode, encode, from_str};
+    use crate::{decode, encode, from_str_flat};
     use std::fs::read_to_string;
 
     #[cfg(feature = "binary")]
     #[test]
     fn test_encode_decode() {
         let test_code = read_to_string("example.bml").unwrap();
-        let stmts = from_str(test_code.as_str()).unwrap();
+        let stmts = from_str_flat(test_code.as_str()).unwrap();
         let encoded = encode(stmts.as_slice());
         assert!(!encoded.is_empty());
         let decoded = decode(encoded.as_slice()).unwrap();