+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.3.0
+
+Features:
+
+* Added config loader construct under feature _loader_
+
+> The Loader builder gives users the ability to load configuration files with more control. It also
+> introduces an ability to load and merge multiple configuration files in a directory
+
+Changes:
+
+* Merged Value and Statement into one enum for easier use
+
 ## 0.2.0 - Binary encoding support
 
 Added feature 'binary' that implements a conversion layer